+++ native/ssz_nif/src/elx_types/beacon_chain.rs
