+++ native/ssz_nif/src/lib.rs
@@ -5,8 +5,8 @@
 //!  - Implement the necessary traits ([`FromElx`] and [`FromLH`]) for its attributes
 //!  - Add the type to [`to_ssz`] and [`from_ssz`] "match" macros
 
-pub(crate) mod lh_types;
-pub(crate) mod types;
+pub(crate) mod elx_types;
+pub(crate) mod rs_types;
 pub(crate) mod utils;
 
 use crate::utils::{helpers::bytes_to_binary, match_schema_and_decode, match_schema_and_encode};
