+++ native/ssz_nif/src/rs_types/beacon_chain.rs
