+++ native/ssz_nif/src/rs_types/mod.rs
@@ -1,0 +1,3 @@
+mod beacon_chain;
+
+pub(crate) use beacon_chain::*;
